+++ solver/src/main/java/org/chocosolver/solver/constraints/ternary/PropXplusYeqZ.java
@@ -58,9 +58,9 @@ public class PropXplusYeqZ extends Propagator<IntVar> {
      * @param Y an integer variable
      * @param Z an integer variable
      */
-    public PropXplusYeqZ(IntVar X, IntVar Y, IntVar Z, boolean enableAC) {
+    public PropXplusYeqZ(IntVar X, IntVar Y, IntVar Z) {
         super(new IntVar[]{X, Y, Z}, PropagatorPriority.TERNARY, false);
-        allbounded = !enableAC || (!X.hasEnumeratedDomain() & !Y.hasEnumeratedDomain() & !Z.hasEnumeratedDomain());
+        allbounded = (!X.hasEnumeratedDomain() & !Y.hasEnumeratedDomain() & !Z.hasEnumeratedDomain());
         r1 = new IntIterableRangeSet();
         r2 = new IntIterableRangeSet();
         r3 = new IntIterableRangeSet();